🔨 Added DashboardItem and ProjectList components
A new file, DashboardItem.tsx, and a new component, ProjectList, were added to the frontend/src/component folder. DashboardItem displays project details in a Card component including the project name, project URL, and number of commits. ProjectList uses useQuery to fetch project data and filters projects by project name.
